refactor: index content
