+++ controllers/system.js
@@ -1,0 +1,4 @@
+
+//save system vars
+const SYSTEM = { }
+module.exports = { SYSTEM }
